修改内容： 调整输入字符的拼接

+++ app/src/main/java/com/example/myapplication/arithmetic/Add.kt
@@ -3,7 +3,7 @@ package com.example.myapplication.arithmetic
 import com.example.myapplication.ui.viewmodle.Record
 
 class Add : Arithmetic {
-    override fun execute(record: Record): Int {
+    override fun execute(record: Record): Double {
         return record.factor1 + record.factor2
     }
 }
