+++ app/src/main/java/com/example/myapplication/arithmetic/Arithmetic.kt
@@ -3,5 +3,5 @@ package com.example.myapplication.arithmetic
 import com.example.myapplication.ui.viewmodle.Record
 
 interface Arithmetic {
-    fun execute(record: Record):Int
+    fun execute(record: Record):Double
 }
