+++ app/src/main/java/com/example/myapplication/arithmetic/EmptyArithmetic.kt
@@ -3,7 +3,7 @@ package com.example.myapplication.arithmetic
 import com.example.myapplication.ui.viewmodle.Record
 
 class EmptyArithmetic : Arithmetic {
-    override fun execute(record: Record): Int {
-        return 0
+    override fun execute(record: Record): Double {
+        return 0.0
     }
 }
