+++ app/src/main/java/com/example/myapplication/arithmetic/Subtract.kt
@@ -3,7 +3,7 @@ package com.example.myapplication.arithmetic
 import com.example.myapplication.ui.viewmodle.Record
 
 class Subtract : Arithmetic {
-    override fun execute(record: Record): Int {
+    override fun execute(record: Record): Double {
         return record.factor1 - record.factor2
     }
 }
